Revised logic for how rc.sabnzbd and sabnzbd.php determines when https is enabled and what port to use in that situation. This commit tries to resolve the following issue;
rc.sabnzbd assumes if https is enabled, then there will be a value in
https_port.  This is not always the case as it is possible for sabnzbd
to have this field blank where it will use HTTPS over the value in regular
port ini entry.  The current logic is not sufficient to account for all
eventualties.

sabnzbd.php has been amended to include additional logic for how to
determine if sabnzbd is running based on whether https is enabled and
what port is in use.

An addtional form field has been contextually added to show when HTTPS is
enabled and does have a specific port value.

There are still some gaps in the logic:

1) php form field HTTPS Port will only show when https is enabled and
contains a value in https_port. It does not take account of when HTTPS is
enabled but value does not exist. In this case SAB is actually running
HTTPS over the HTTP port, but the unraid sab webgui will only show
"HTTP Port" which could lead to some user confusion. Perhaps it is best
to hide the HTTPS port value at all times. Or add an extra contextual
change so that "HTTP Port" form field becomes "HTTPS Port" when https is
enabled but https_port has no value.

2) Changing the HTTPS / PORT from SAB webgui config screen can lead to
confusion in the plugin webgui. The user always has the option to make
changes within SABNZBD config and hit the built-in restart. For HTTPS and
the port, this can cause the rc.script and php file to get out of sync
such as the script may try to stop an existing instance of SAB but is
looking for the wrong PID-port file. PHP file also wrongly detects SAB as
not running.