Update couchpotato_v2_unplugged.plg
The location of the PORT value that the plugin greps for has moved in the settings.conf file.  With 'awk 'NR==2'', the port value is not applied correctly on a restart, but with 'awk 'NR==1'', it is.  

I simply changed this in the three places it is relevant in the plugin file.

More info here:
http://lime-technology.com/forum/index.php?topic=21260.msg292453#msg292453